About
Nowadays every platform has reports and most of the time we need to convert the files into PDF formats. 
I got a similar task a few days back. I have also done this conversion to PDF, but this time I decided to explore a serverless approach. 
We are going to explore with Lambda. Lambda is an event-driven,  serverless computing service that we can integrate with many 
other services like S3, SNS, DynamoDB, etc.